Feat: 사용하지 않는 showEraserView 삭제
